Move types not directly exposed in routes to *Handler.types.ts files
