+++ app/components/LicensePage.tsx
@@ -5,7 +5,7 @@ import { HeaderBar } from 'cloudify-ui-components';
 import type { ButtonProps } from 'semantic-ui-react';
 import styled from 'styled-components';
 
-import type { LicenseResponse } from '../../backend/routes/Auth.types';
+import type { LicenseResponse } from '../../backend/handler/AuthHandler.types';
 import type { LicenseStatus } from '../reducers/managerReducer/licenseReducer';
 import Banner from './banner/Banner';
 import Consts from '../utils/consts';
