+++ app/reducers/managerReducer/licenseReducer.ts
@@ -2,7 +2,7 @@ import type { Reducer } from 'redux';
 import * as types from '../../actions/types';
 import Auth from '../../utils/auth';
 import type Consts from '../../utils/consts';
-import type { LicenseResponse } from '../../../backend/routes/Auth.types';
+import type { LicenseResponse } from '../../../backend/handler/AuthHandler.types';
 
 type ValueOf<T> = T[keyof T];
 export type LicenseStatus = ValueOf<typeof Consts.LICENSE>;
