+++ app/utils/auth.ts
@@ -6,7 +6,8 @@ import External from './External';
 import Internal from './Internal';
 import encodeTextToBase64 from './encodeTextToBase64';
 import type { ManagerData, LicenseData, LicenseStatus } from '../reducers/managerReducer';
-import type { GetAuthUserResponse, LicenseResponse } from '../../backend/routes/Auth.types';
+import type { GetAuthUserResponse } from '../../backend/routes/Auth.types';
+import type { LicenseResponse } from '../../backend/handler/AuthHandler.types';
 
 export default class Auth {
     static login(username: string, password: string) {
