+++ backend/handler/FilterHandler.types.ts
@@ -1,0 +1,5 @@
+export interface FilterUses {
+    pageName: string;
+    widgetName: string;
+    username: string;
+}
