+++ backend/handler/SourceHandler.types.ts
@@ -1,0 +1,6 @@
+export interface ScanningItem {
+    key: string;
+    title: string;
+    isDir: boolean;
+    children?: ScanningItem[];
+}
