+++ backend/handler/WidgetsHandler.types.ts
@@ -1,0 +1,9 @@
+export interface WidgetData {
+    id: string;
+    isCustom: boolean;
+}
+
+export interface WidgetUsage {
+    username: string;
+    managerIp: string;
+}
