+++ backend/routes/Templates.types.ts
@@ -1,98 +1,20 @@
-export interface WidgetDefinition {
-    definition: string;
-    name: string;
-    height: number;
-    width: number;
-    x: number;
-    y: number;
-    configuration: Record<string, any>;
-}
-
-export interface WidgetsSection {
-    type: 'widgets';
-    content: WidgetDefinition[];
-}
-
-export interface TabContent {
-    name: string;
-    widgets: WidgetDefinition[];
-    isDefault?: boolean;
-}
-
-export interface TabsSection {
-    type: 'tabs';
-    content: TabContent[];
-}
-
-export type LayoutSectionType = 'tabs' | 'widgets';
-
-export type LayoutSection = WidgetsSection | TabsSection;
-
-export interface PageItem {
-    id: string;
-    type: 'page' | 'pageGroup';
-}
-
-export interface CommonFileContent {
-    updatedBy?: string;
-    updatedAt?: string;
-}
-
-export interface CommonIdentityData {
-    id: string;
-    name: string;
-    custom: boolean;
-}
-
-export interface TemplateFileContent extends CommonFileContent {
-    roles: string[];
-    tenants: string[];
-    pages: PageItem[];
-}
-
-export interface TemplateData {
-    roles: TemplateFileContent['roles'];
-    tenants: TemplateFileContent['tenants'];
-}
-
-export interface Template extends Required<CommonFileContent>, CommonIdentityData {
-    data: TemplateData & {
-        pages: TemplateFileContent['pages'];
-    };
-}
-
-export interface PageGroupFileContent extends CommonFileContent {
-    name: string;
-    icon: string;
-    pages: string[];
-}
-
-export type PageGroup = Required<PageGroupFileContent>;
-
-export interface PageFileContent extends CommonFileContent {
-    name: string;
-    icon?: string;
-    layout: LayoutSection[];
-}
-
-export interface Page extends Required<CommonFileContent>, CommonIdentityData {
-    data: {
-        icon: PageFileContent['icon'];
-        layout: PageFileContent['layout'];
-    };
-}
+import type {
+    CreatePageData,
+    CreatePageGroupData,
+    CreateTemplateData,
+    Page,
+    PageGroup,
+    Template,
+    UpdatePageData,
+    UpdatePageGroupData,
+    UpdateTemplateData
+} from '../handler/templates/types';
 
 export type GetTemplatesResponse = Template[];
 
-export interface PostTemplatesRequestBody {
-    id: string;
-    data: TemplateData;
-    pages: PageItem[];
-}
+export type PostTemplatesRequestBody = CreateTemplateData;
 
-export interface PutTemplatesRequestBody extends PostTemplatesRequestBody {
-    oldId: string;
-}
+export type PutTemplatesRequestBody = UpdateTemplateData;
 
 export type GetSelectTemplateResponse = string;
 
@@ -102,25 +24,12 @@ export interface GetSelectTemplateQueryParams {
 
 export type GetPageGroupsResponse = PageGroup[];
 
-export interface PostPageGroupsRequestBody {
-    id: string;
-    icon: string;
-    name: string;
-    pages: string[];
-}
+export type PostPageGroupsRequestBody = CreatePageGroupData;
 
-export type PutPageGroupsRequestBody = PostPageGroupsRequestBody;
+export type PutPageGroupsRequestBody = UpdatePageGroupData;
 
 export type GetPagesResponse = (Page | null)[];
 
-export interface PostPagesRequestBody {
-    id: string;
-    name: string;
-    layout: [];
-}
+export type PostPagesRequestBody = CreatePageData;
 
-export interface PutPagesRequestBody extends Omit<PostPagesRequestBody, 'layout'> {
-    oldId: string;
-    icon?: string;
-    layout: LayoutSection[];
-}
+export type PutPagesRequestBody = UpdatePageData;
